rewrote envelope

+++ src/channel.h
@@ -11,7 +11,7 @@ class channel
 public:
     channel(int samplerate=44100, double volume=1.0);
     ~channel();
-    short get(double time, int waveform, double hertz);
+    double get(double time, int waveform, double hertz);
     double getSine(double time, double hertz);
     double getTriangle(double time, double hertz);
     double getSquare(double time, double hertz);
